Extracted cameraInspector from components.js
